Adjust person licence application repository test class name
